fix quoting when passing string to shell `eval`
The argument to `eval` is enclosed in double quotes and double quotes
within the string are escaped. As any dollar signs in the string are
not escaped they will be subject to parameter expansion before the string
is evaluated. This means that code such as

    define-command peneira-git-files %{
        peneira 'files: ' %{
            toplevel="$(git -C "${kak_buffile%/*}" rev-parse --show-toplevel 2>/dev/null)"
            pretty_toplevel="~/${toplevel#$HOME/}/"
            git -C "$toplevel" -c core.quotePath=false ls-files | sed "s|^|$pretty_toplevel|"
        }
    }

Does not work because instead of running

    git -C "/path/to/dir" rev-parse ...

it runs

    git -C \"/path/to/dir\" rev-parse ...

In order to work correctly every '$' has to be escaped with a '\' which
is unexpected and rather confusing.

The user's code is also run in the main shell process which means that
if it calls `exit` peneira shows no output because the main shell exits
early. Fix these two issues by correctly quoting the user's code before
passing it to `eval` to suppress parameter expansion and running the
code in a subshell so that calling `exit` will exit the subshell and
not the main shell. Using a subshell also prevents the user's code from
accidentally overwriting shell variables used by peneira.

As this is a breaking change the buffer list example in the documentation
that relies on the confusing behavior of expanding variables before
passing them to `eval` is updated to accommodate the quoting changes.